pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.30",
    "ollama>=0.6.0",
    "pypdf>=6.1.1",
    "python-docx>=1.2.0",
    "werkzeug>=3.1.3",
]
